NO MEMORY ERROR FINALLY
Il reste cependant un detail, les message LROA renvoyer depuis le LORA sont anormalement élevée et ne se reduisent pas en debranchant l'antenne ou en relancant le debug ou en reset hadware la carte

+++ BASELIQ/Core/Inc/shared.h
@@ -72,7 +72,7 @@ typedef struct{
 	uint8_t recipient;
 	uint8_t sender;
 	uint8_t type;
-	uint8_t len_payload;
+	size_t len_payload;
 }LORA_Header;
 
 typedef struct {
